Fix Reference Cycle (#255)
### Description

Fixes a strong reference cycle in `TextViewController` causing it to not
release when removed from the view heirarchy.

### Related Issues

* https://github.com/CodeEditApp/CodeEdit/issues/1794

### Checklist

<!--- Add things that are not yet implemented above -->

- [x] I read and understood the [contributing
guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md)
as well as the [code of
conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots

Stable memory usage in the example app (mem jumps are opening & closing
a large file a few times):
![Screenshot 2024-07-03 at 8 22
20 PM](https://github.com/CodeEditApp/CodeEditSourceEditor/assets/35942988/7532e0f5-f72e-44b5-ac8c-5f194736d3d4)
