+++ Sources/CodeEditSourceEditor/Controller/TextViewController+ToggleComment.swift
@@ -12,7 +12,6 @@ extension TextViewController {
     /// Method called when CMD + / key sequence recognized, comments cursor's current line of code
     public func commandSlashCalled() {
         guard let cursorPosition = cursorPositions.first else {
-            print("There is no cursor \(#function)")
             return
         }
         // Many languages require a character sequence at the beginning of the line to comment the line.
@@ -33,12 +32,8 @@ extension TextViewController {
 
     ///  Toggles comment string at the beginning of a specified line (lineNumber is 1-indexed)
     private func toggleCharsAtBeginningOfLine(chars: String, lineNumber: Int) {
-        guard let lineInfo = textView.layoutManager.textLineForIndex(lineNumber - 1) else {
-            print("There are no characters/lineInfo \(#function)")
-            return
-        }
-        guard let lineString = textView.textStorage.substring(from: lineInfo.range) else {
-            print("There are no characters/lineString \(#function)")
+        guard let lineInfo = textView.layoutManager.textLineForIndex(lineNumber - 1),
+              let lineString = textView.textStorage.substring(from: lineInfo.range) else {
             return
         }
         let firstNonWhiteSpaceCharIndex = lineString.firstIndex(where: {!$0.isWhitespace}) ?? lineString.startIndex
@@ -46,28 +41,22 @@ extension TextViewController {
         let firstCharsInLine = lineString.suffix(from: firstNonWhiteSpaceCharIndex).prefix(chars.count)
         // toggle comment off
         if firstCharsInLine == chars {
-            textView.replaceCharacters(in: NSRange(
-                location: lineInfo.range.location + numWhitespaceChars,
-                length: chars.count
-            ), with: "")
-        }
-        // toggle comment on
-        else {
-            textView.replaceCharacters(in: NSRange(
-                location: lineInfo.range.location + numWhitespaceChars,
-                length: 0
-            ), with: chars)
+            textView.replaceCharacters(
+                in: NSRange(location: lineInfo.range.location + numWhitespaceChars, length: chars.count),
+                with: ""
+            )
+        } else {
+            // toggle comment on
+            textView.replaceCharacters(
+                in: NSRange(location: lineInfo.range.location + numWhitespaceChars, length: 0),
+                with: chars
+            )
         }
     }
 
     ///  Toggles a specific string of characters at the end of a specified line. (lineNumber is 1-indexed)
     private func toggleCharsAtEndOfLine(chars: String, lineNumber: Int) {
-        guard let lineInfo = textView.layoutManager.textLineForIndex(lineNumber - 1) else {
-            print("There are no characters/lineInfo \(#function)")
-            return
-        }
-        guard let lineString = textView.textStorage.substring(from: lineInfo.range) else {
-            print("There are no characters/lineString \(#function)")
+        guard let lineInfo = textView.layoutManager.textLineForIndex(lineNumber - 1), !lineInfo.range.isEmpty else {
             return
         }
         let lineLastCharIndex = lineInfo.range.location + lineInfo.range.length - 1
@@ -79,13 +68,12 @@ extension TextViewController {
         let lastCharsInLine = textView.textStorage.substring(from: closeCommentRange)
         // toggle comment off
         if lastCharsInLine == chars {
-            textView.replaceCharacters(in: NSRange(
-                location: lineLastCharIndex - closeCommentLength,
-                length: closeCommentLength
-            ), with: "")
-        }
-        // toggle comment on
-        else {
+            textView.replaceCharacters(
+                in: NSRange(location: lineLastCharIndex - closeCommentLength, length: closeCommentLength),
+                with: ""
+            )
+        } else {
+            // toggle comment on
             textView.replaceCharacters(in: NSRange(location: lineLastCharIndex, length: 0), with: chars)
         }
     }
