+++ Sources/CodeEditSourceEditor/Controller/TextViewController.swift
@@ -28,6 +28,7 @@ public class TextViewController: NSViewController {
     internal var highlightLayers: [CALayer] = []
     internal var systemAppearance: NSAppearance.Name?
 
+    package var localEvenMonitor: Any?
     package var isPostingCursorNotification: Bool = false
 
     /// The string contents.
@@ -172,7 +173,7 @@ public class TextViewController: NSViewController {
     /// This will be `nil` if another highlighter provider is passed to the source editor.
     internal(set) public var treeSitterClient: TreeSitterClient?
 
-    private var fontCharWidth: CGFloat { (" " as NSString).size(withAttributes: [.font: font]).width }
+    package var fontCharWidth: CGFloat { (" " as NSString).size(withAttributes: [.font: font]).width }
 
     /// Filters used when applying edits..
     internal var textFilters: [TextFormation.Filter] = []
@@ -180,7 +181,7 @@ public class TextViewController: NSViewController {
     internal var cancellables = Set<AnyCancellable>()
 
     /// ScrollView's bottom inset using as editor overscroll
-    private var bottomContentInsets: CGFloat {
+    package var bottomContentInsets: CGFloat {
         let height = view.frame.height
         var inset = editorOverscroll * height
 
@@ -270,15 +271,7 @@ public class TextViewController: NSViewController {
     // MARK: Paragraph Style
 
     /// A default `NSParagraphStyle` with a set `lineHeight`
-    internal lazy var paragraphStyle: NSMutableParagraphStyle = generateParagraphStyle()
-
-    private func generateParagraphStyle() -> NSMutableParagraphStyle {
-        // swiftlint:disable:next force_cast
-        let paragraph = NSParagraphStyle.default.mutableCopy() as! NSMutableParagraphStyle
-        paragraph.tabStops.removeAll()
-        paragraph.defaultTabInterval = CGFloat(tabWidth) * fontCharWidth
-        return paragraph
-    }
+    package lazy var paragraphStyle: NSMutableParagraphStyle = generateParagraphStyle()
 
     // MARK: - Reload UI
 
@@ -293,59 +286,6 @@ public class TextViewController: NSViewController {
         highlighter?.invalidate()
     }
 
-    /// Style the text view.
-    package func styleTextView() {
-        textView.selectionManager.selectionBackgroundColor = theme.selection
-        textView.selectionManager.selectedLineBackgroundColor = getThemeBackground()
-        textView.selectionManager.highlightSelectedLine = isEditable
-        textView.selectionManager.insertionPointColor = theme.insertionPoint
-        paragraphStyle = generateParagraphStyle()
-        textView.typingAttributes = attributesFor(nil)
-    }
-
-    /// Finds the preferred use theme background.
-    /// - Returns: The background color to use.
-    private func getThemeBackground() -> NSColor {
-        if useThemeBackground {
-            return theme.lineHighlight
-        }
-
-        if systemAppearance == .darkAqua {
-            return NSColor.quaternaryLabelColor
-        }
-
-        return NSColor.selectedTextBackgroundColor.withSystemEffect(.disabled)
-    }
-
-    /// Style the gutter view.
-    package func styleGutterView() {
-        gutterView.frame.origin.y = -scrollView.contentInsets.top
-        gutterView.selectedLineColor = useThemeBackground ? theme.lineHighlight : systemAppearance == .darkAqua
-        ? NSColor.quaternaryLabelColor
-        : NSColor.selectedTextBackgroundColor.withSystemEffect(.disabled)
-        gutterView.highlightSelectedLines = isEditable
-        gutterView.font = font.rulerFont
-        gutterView.backgroundColor = useThemeBackground ? theme.background : .textBackgroundColor
-        if self.isEditable == false {
-            gutterView.selectedLineTextColor = nil
-            gutterView.selectedLineColor = .clear
-        }
-    }
-
-    /// Style the scroll view.
-    package func styleScrollView() {
-        guard let scrollView = view as? NSScrollView else { return }
-        scrollView.drawsBackground = useThemeBackground
-        scrollView.backgroundColor = useThemeBackground ? theme.background : .clear
-        if let contentInsets {
-            scrollView.automaticallyAdjustsContentInsets = false
-            scrollView.contentInsets = contentInsets
-        } else {
-            scrollView.automaticallyAdjustsContentInsets = true
-        }
-        scrollView.contentInsets.bottom = (contentInsets?.bottom ?? 0) + bottomContentInsets
-    }
-
     deinit {
         if let highlighter {
             textView.removeStorageDelegate(highlighter)
@@ -354,6 +294,10 @@ public class TextViewController: NSViewController {
         highlightProvider = nil
         NotificationCenter.default.removeObserver(self)
         cancellables.forEach { $0.cancel() }
+        if let localEvenMonitor {
+            NSEvent.removeMonitor(localEvenMonitor)
+        }
+        localEvenMonitor = nil
     }
 }
